Refactored and added tie breaking rule in sorting and ability to ignore the root note to simplify the chord name
Added more interval macros and existing macro references are changed to reflect more accurate intervals.

At least two notes are required in the input.

Whenever the ranking of two chord names are same during sorting, the lenghts of the chord names are used for tie breaking.

Ability to output a cleaner chord name whenever the lowest (root) note can be omitted in a slashed chord. For example, F/G for {G, F, A, C} but Fadd9/G for {G, F, A, C, G}